pyproject.toml
[tool.poetry]
name = "hhSCNet"
version = "0.1.0"
description = ""
authors = ["Hunter Hogan <HunterHogan@pm.me>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
numpy = "^2.2.0"
ml-collections = "^1.0.0"
tqdm = "^4.67.1"
accelerate = "^1.2.1"
einops = "^0.8.0"
julius = "^0.2.7"
soundfile = "^0.12.1"
torch = "^2.5.1"
torchaudio = "^2.5.1"
ruamel-yaml = "^0.18.6"

[tool.poetry.scripts]
hhSCNet = "hhSCNet:main"

[tool.poetry.group.test.dependencies]
pytest = "^8.3.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
